goit-nodejs-hw-05-03

+++ src/services/contacts.js
@@ -2,14 +2,16 @@ import { ContactsCollection } from '../db/models/contact.js';
 import { calculatePaginationData } from '../utils/calculatePaginationData.js';
 import { SORT_ORDER } from "../constants/index.js";
 
-export const getAllContacts = async ({ page = 1,
+export const getAllContacts = async ({
+  userId,
+  page = 1,
   perPage = 10,
   sortOrder = SORT_ORDER.ASC,
   sortBy = '_id',
   filter = {}, }) => {
   const limit = perPage;
   const skip = (page - 1) * perPage;
-  const contactsQuery = ContactsCollection.find();
+  const contactsQuery = ContactsCollection.find({userId});
 
 if (filter.contactType) {
     contactsQuery.where('contactType').equals(filter.contactType);
@@ -19,7 +21,7 @@ if (filter.isFavourite !== undefined) {
 }
 
 const [contactsCount, contacts] = await Promise.all([
-    ContactsCollection.find().merge(contactsQuery).countDocuments(),
+    ContactsCollection.find({userId}).merge(contactsQuery).countDocuments(),
     contactsQuery
       .skip(skip)
       .limit(limit)
@@ -35,30 +37,26 @@ const [contactsCount, contacts] = await Promise.all([
   };
 };
 
-export const getContactById = async (contactId) => {
-  const contact = await ContactsCollection.findById(contactId);
+export const getContactById = async (id, userId) => {
+  const contact = await ContactsCollection.findOne({_id: id, userId});
   return contact;
 };
 
-export const createContact = async (payload) => {
-  const contact = await ContactsCollection.create(payload);
+export const createContact = async (payload, userId) => {
+  const contact = await ContactsCollection.create(payload, userId);
   return contact;
 };
 
-export const deleteContact = async (contactId) => {
-  const contact = await ContactsCollection.findOneAndDelete({
-    _id: contactId,
-  });
+export const deleteContact = async (id, userId) => {
+  const contact = await ContactsCollection.findOneAndDelete({_id: id, userId});
 
   return contact;
 };
 
-export const updateContact = async (contactId, payload, options = {}) => {
-  const rawResult = await ContactsCollection.findOneAndUpdate(
-    { _id: contactId },
+export const updateContact = async (id, payload, userId, options = {}) => {
+  const rawResult = await ContactsCollection.findOneAndUpdate({ _id: id, userId},
     payload,
-    {
-      new: true,
+    { new: true,
       includeResultMetadata: true,
       ...options,
     },
